requirements.txt
torch
transformers
pandas
scikit-learn
pyyaml
safetensors